sidebar more tweaks
